requirements.txt
pandas>=2.2,<3.0
snowflake-connector-python[pandas]>=3.11,<4.0
python-dotenv>=1.0,<2.0

dbt-core>=1.8,<2.0
dbt-snowflake>=1.8,<2.0
dbt-duckdb>=1.8,<2.0
dagster>=1.12,<2.0
dagster-cloud>=1.12,<2.0

ruff>=0.6,<1.0
sqlfluff>=3.1,<4.0
sqlfluff-templater-dbt>=3.1,<4.0
pre-commit>=3.7,<4.0
